Renamed PhotonNumberConvergenceChecker -> ChiSquaredPhotonNumberConvergenceChecker. Created new PhotonNumberConvergenceChecker interface and factory. Created PassivePhotonNumberConvergenceChecker.
